requirements.txt
django==5.1.1
djangorestframework==3.15.2
requests==2.22.0
pytz==2020.1
django-slack==5.14.3
SQLAlchemy==1.3.13
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
drf-nested-routers==0.91
django-cors-headers
PyJWT
